simplified track-manager organization
