--- applications/track-manager/src/main/scala/com/hashmapinc/tempus/edge/track_manager/db/DbConnection.scala
@@ -1,7 +1,0 @@
-package com.hashmapinc.tempus.edge.track_manager.opc
-
-import com.typesafe.scalalogging.Logger
-
-object DbConnection {
-  private val logger = Logger(getClass())
-}
