--- applications/track-manager/src/main/scala/com/hashmapinc/tempus/edge/track_manager/db/DbController.scala
@@ -1,8 +1,0 @@
-package com.hashmapinc.tempus.edge.track_manager.db
-
-import com.typesafe.scalalogging.Logger
-
-object DbController {
-  private val logger = Logger(getClass())
-
-}
