--- applications/track-manager/src/test/scala/com/hashmapinc/tempus/edge/track_manager/db/DbControllerTest.scala
@@ -1,7 +1,0 @@
-package com.hashmapinc.tempus.edge.track_manager.db
-
-import org.scalatest.FlatSpec
-
-class DbControllerTest extends FlatSpec {
-  //"IofogController" should ""
-}
