+++ applications/track-manager/src/test/scala/com/hashmapinc/tempus/edge/track_manager/iofog/IofogControllerTest.scala
@@ -3,5 +3,4 @@ package com.hashmapinc.tempus.edge.track_manager.iofog
 import org.scalatest.FlatSpec
 
 class IofogControllerTest extends FlatSpec {
-  //"IofogController" should ""
 }
